modernize - because refactored implementations do not rely on dependencies, the unit tests can be redesigned as such
